Shim for svg elements
svg elements fail to be created properly in AngularJS, because it
appears the elements are not created under the correct namespace.

[related SO discussions]

This prevents svg elements from being used in AngularJS directives
which is an exceedingly powerful means for building new primitives
and thereby abstracting complexity.

The solution is to manually create svg elements and insert them
into the DOM. In general, this means including a custom compile step
for any directive that creates svg elements, such as those with svg
elements in their templates.

This seemed like a pretty intrusive pattern. For one, when the
framework gets fixed, all the compile steps would need to be removed.
For another, it's a detail of the framework that ideally should be
abstracted from the directive programmer, who already has a lot to
think about.

The proxy pattern solves this issue by standing in between the
element that is actually used in programming the directive, and the
raw DOM element that is created behind the scene. It utilizes a
proxy directive that orchestrates transmitting attribute changes
from the proxy element to the raw DOM where it's interpretted.

There may be other powerful ways to use the proxy pattern which I
might explore later.

To use the pattern, the following is needed:

1. Add AngularJS directives for all svg-namespace directives
   replacing raw elements for one created in the proper namespace.
2. Use the svg element names (to which AngularJS will apply the
   directives defined above) in templates as you would use any
   other html directive.
3. Add the "proxy" directive to svg elements used in any
   template.

Now, when the framework is fixed, we only need to remove the shim,
and the proxy directive wherever svg elements are used.

See app/views/main.html for a simple demo.